Upgrade to .NET 9

+++ StravaRaceAPI/Services/ErrorMessages.cs
@@ -5,10 +5,10 @@ namespace StravaRaceAPI.Services;
 /// </summary>
 public static class ErrorMessages
 {
-    private static readonly Func<string, int, string> NotFoundMessage =
+    private static readonly Func<string, ulong, string> NotFoundMessage =
         (item, userId) => $"{item} with id {userId} not found.";
 
-    public static readonly Func<int, string> UserNotFoundMessage = userId => NotFoundMessage("User", userId);
+    public static readonly Func<int, string> UserNotFoundMessage = userId => NotFoundMessage("User", (ulong)userId);
 
-    public static readonly Func<int, string> EventNotFoundMessage = eventId => NotFoundMessage("Event", eventId);
+    public static readonly Func<ulong, string> EventNotFoundMessage = eventId => NotFoundMessage("Event", eventId);
 }
